#58: Moved TokenDefinitions to test folder.
