+++ core/src/test/java/io/parsingdata/metal/BackTrackNameBindTest.java
@@ -19,10 +19,10 @@ package io.parsingdata.metal;
 import static io.parsingdata.metal.Shorthand.cho;
 import static io.parsingdata.metal.Shorthand.rep;
 import static io.parsingdata.metal.Shorthand.seq;
-import static io.parsingdata.metal.TokenDefinitions.any;
-import static io.parsingdata.metal.TokenDefinitions.eq;
-import static io.parsingdata.metal.TokenDefinitions.eqRef;
-import static io.parsingdata.metal.TokenDefinitions.notEqRef;
+import static io.parsingdata.metal.util.TokenDefinitions.any;
+import static io.parsingdata.metal.util.TokenDefinitions.eq;
+import static io.parsingdata.metal.util.TokenDefinitions.eqRef;
+import static io.parsingdata.metal.util.TokenDefinitions.notEqRef;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
