+++ core/src/test/java/io/parsingdata/metal/ReadUntilTest.java
@@ -18,8 +18,8 @@ package io.parsingdata.metal;
 
 import static io.parsingdata.metal.Shorthand.rep;
 import static io.parsingdata.metal.Shorthand.seq;
-import static io.parsingdata.metal.TokenDefinitions.eq;
-import static io.parsingdata.metal.TokenDefinitions.notEq;
+import static io.parsingdata.metal.util.TokenDefinitions.eq;
+import static io.parsingdata.metal.util.TokenDefinitions.notEq;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
