+++ core/src/test/java/io/parsingdata/metal/ReferenceValueExpressionSemanticsTest.java
@@ -28,8 +28,8 @@ import java.util.Arrays;
 import java.util.Collection;
 
 import static io.parsingdata.metal.Shorthand.*;
-import static io.parsingdata.metal.TokenDefinitions.any;
-import static io.parsingdata.metal.TokenDefinitions.eqRef;
+import static io.parsingdata.metal.util.TokenDefinitions.any;
+import static io.parsingdata.metal.util.TokenDefinitions.eqRef;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
