+++ core/src/test/java/io/parsingdata/metal/expression/value/ElvisExpressionTest.java
@@ -25,7 +25,7 @@ import org.junit.Test;
 import java.io.IOException;
 
 import static io.parsingdata.metal.Shorthand.*;
-import static io.parsingdata.metal.TokenDefinitions.any;
+import static io.parsingdata.metal.util.TokenDefinitions.any;
 import static io.parsingdata.metal.util.EncodingFactory.enc;
 import static io.parsingdata.metal.util.EnvironmentFactory.stream;
 import static org.hamcrest.CoreMatchers.equalTo;
